Extract LLMS text helpers
